refactor: remove tailwind prefix in rule file names
