+++ src/api/defaults.ts
@@ -4,7 +4,7 @@ import {
   DEFAULT_TAG_NAMES,
   DEFAULT_VARIABLE_NAMES
 } from "better-tailwindcss:options:default-options.js";
-import { DEFAULT_IGNORED_UNREGISTERED_CLASSES } from "better-tailwindcss:rules:tailwind-no-unregistered-classes.js";
+import { DEFAULT_IGNORED_UNREGISTERED_CLASSES } from "better-tailwindcss:rules:no-unregistered-classes.js";
 
 
 export function getDefaultCallees() {
