+++ src/configs/config.ts
@@ -1,10 +1,10 @@
-import { multiline } from "better-tailwindcss:rules:tailwind-multiline.js";
-import { noConflictingClasses } from "better-tailwindcss:rules:tailwind-no-conflicting-classes.js";
-import { noDuplicateClasses } from "better-tailwindcss:rules:tailwind-no-duplicate-classes.js";
-import { noRestrictedClasses } from "better-tailwindcss:rules:tailwind-no-restricted-classes.js";
-import { noUnnecessaryWhitespace } from "better-tailwindcss:rules:tailwind-no-unnecessary-whitespace.js";
-import { noUnregisteredClasses } from "better-tailwindcss:rules:tailwind-no-unregistered-classes.js";
-import { sortClasses } from "better-tailwindcss:rules:tailwind-sort-classes.js";
+import { multiline } from "better-tailwindcss:rules:multiline.js";
+import { noConflictingClasses } from "better-tailwindcss:rules:no-conflicting-classes.js";
+import { noDuplicateClasses } from "better-tailwindcss:rules:no-duplicate-classes.js";
+import { noRestrictedClasses } from "better-tailwindcss:rules:no-restricted-classes.js";
+import { noUnnecessaryWhitespace } from "better-tailwindcss:rules:no-unnecessary-whitespace.js";
+import { noUnregisteredClasses } from "better-tailwindcss:rules:no-unregistered-classes.js";
+import { sortClasses } from "better-tailwindcss:rules:sort-classes.js";
 
 import type { ESLint } from "eslint";
