+++ src/options/callees/clb.test.ts
@@ -5,7 +5,7 @@ import {
   CLB_COMPOUND_VARIANTS_CLASSES,
   CLB_VARIANT_VALUES
 } from "better-tailwindcss:options:callees/clb.js";
-import { noUnnecessaryWhitespace } from "better-tailwindcss:rules:tailwind-no-unnecessary-whitespace.js";
+import { noUnnecessaryWhitespace } from "better-tailwindcss:rules:no-unnecessary-whitespace.js";
 import { lint, TEST_SYNTAXES } from "better-tailwindcss:tests:utils.js";
