+++ src/options/callees/clsx.test.ts
@@ -1,7 +1,7 @@
 import { describe, it } from "vitest";
 
 import { CLSX_OBJECT_KEYS, CLSX_STRINGS } from "better-tailwindcss:options:callees/clsx.js";
-import { noUnnecessaryWhitespace } from "better-tailwindcss:rules:tailwind-no-unnecessary-whitespace.js";
+import { noUnnecessaryWhitespace } from "better-tailwindcss:rules:no-unnecessary-whitespace.js";
 import { lint, TEST_SYNTAXES } from "better-tailwindcss:tests:utils.js";
