+++ src/options/callees/cva.test.ts
@@ -5,7 +5,7 @@ import {
   CVA_STRINGS,
   CVA_VARIANT_VALUES
 } from "better-tailwindcss:options:callees/cva.js";
-import { noUnnecessaryWhitespace } from "better-tailwindcss:rules:tailwind-no-unnecessary-whitespace.js";
+import { noUnnecessaryWhitespace } from "better-tailwindcss:rules:no-unnecessary-whitespace.js";
 import { lint, TEST_SYNTAXES } from "better-tailwindcss:tests:utils.js";
