+++ src/options/callees/objstr.test.ts
@@ -1,7 +1,7 @@
 import { describe, it } from "vitest";
 
 import { OBJSTR_OBJECT_KEYS } from "better-tailwindcss:options:callees/objstr.js";
-import { noUnnecessaryWhitespace } from "better-tailwindcss:rules:tailwind-no-unnecessary-whitespace.js";
+import { noUnnecessaryWhitespace } from "better-tailwindcss:rules:no-unnecessary-whitespace.js";
 import { lint, TEST_SYNTAXES } from "better-tailwindcss:tests:utils.js";
