+++ src/rules/multiline.test.ts
@@ -1,6 +1,6 @@
 import { describe, it } from "vitest";
 
-import { multiline } from "better-tailwindcss:rules:tailwind-multiline.js";
+import { multiline } from "better-tailwindcss:rules:multiline.js";
 import { createTrimTag, lint, TEST_SYNTAXES } from "better-tailwindcss:tests:utils.js";
 import { MatcherType } from "better-tailwindcss:types:rule.js";
