+++ src/rules/no-conflicting-classes.ts
