+++ src/rules/no-duplicate-classes.test.ts
@@ -1,6 +1,6 @@
 import { describe, it } from "vitest";
 
-import { noDuplicateClasses } from "better-tailwindcss:rules:tailwind-no-duplicate-classes.js";
+import { noDuplicateClasses } from "better-tailwindcss:rules:no-duplicate-classes.js";
 import { createTrimTag, lint, TEST_SYNTAXES } from "better-tailwindcss:tests:utils.js";
