+++ src/rules/no-duplicate-classes.ts
