+++ src/rules/no-restricted-classes.ts
