+++ src/rules/no-unnecessary-whitespace.ts
