+++ src/rules/no-unregistered-classes.test.ts
@@ -1,7 +1,7 @@
 import { getTailwindcssVersion, TailwindcssVersion } from "src/tailwind/utils/version.js";
 import { describe, it } from "vitest";
 
-import { noUnregisteredClasses } from "better-tailwindcss:rules:tailwind-no-unregistered-classes.js";
+import { noUnregisteredClasses } from "better-tailwindcss:rules:no-unregistered-classes.js";
 import { lint, TEST_SYNTAXES } from "better-tailwindcss:tests:utils.js";
