+++ src/rules/no-unregistered-classes.ts
